feat: 抽象出终端为独立组件
